requirements.txt
pandas
numpy
opencv-python
PyYAML
roboflow
ultralytics==8.0.33
scikit-learn==0.21.0
tensorflow==2.11.0
scikit-image==0.19.3
filterpy==1.4.5
jupyter
